# server/src/middleware/__generated__/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ReagentTradingType": {
				"type": "string",
				"enum": [
					"trade",
					"giveaway",
					"sell"
				]
			},
			"ReagentCategory": {
				"type": "string",
				"enum": [
					"chemical",
					"hazardous",
					"biological"
				]
			},
			"ReagentVisibility": {
				"type": "string",
				"enum": [
					"everyone",
					"region",
					"institution",
					"private"
				]
			},
			"Reagent": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"condition": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"expiryDate": {
						"type": "string"
					},
					"tradingType": {
						"$ref": "#/components/schemas/ReagentTradingType"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/ReagentCategory"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"createdAtReadable": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"unit": {
						"type": "string"
					},
					"visibility": {
						"$ref": "#/components/schemas/ReagentVisibility"
					},
					"requesterOfferedReagentId": {
						"type": "string"
					},
					"restricted": {
						"type": "boolean"
					}
				},
				"required": [
					"user_id",
					"name",
					"description",
					"condition",
					"quantity",
					"expiryDate",
					"tradingType",
					"categories",
					"createdAt",
					"createdAtReadable",
					"location",
					"unit",
					"restricted"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"User": {
				"properties": {
					"email": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"university": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"about": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"lab_manager",
							"admin"
						]
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"displayName",
					"university",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_User_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"university": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"about": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"lab_manager",
							"admin"
						]
					},
					"image": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"CreateReagentRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"condition": {
						"type": "string"
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/ReagentCategory"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unit": {
						"type": "string"
					},
					"expiryDate": {
						"type": "string"
					},
					"tradingType": {
						"$ref": "#/components/schemas/ReagentTradingType"
					},
					"location": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"visibility": {
						"$ref": "#/components/schemas/ReagentVisibility"
					},
					"restricted": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"condition",
					"categories",
					"quantity",
					"unit",
					"expiryDate",
					"tradingType",
					"location",
					"restricted"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Reagent_": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"condition": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"expiryDate": {
						"type": "string"
					},
					"tradingType": {
						"$ref": "#/components/schemas/ReagentTradingType"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"categories": {
						"items": {
							"$ref": "#/components/schemas/ReagentCategory"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"createdAtReadable": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"unit": {
						"type": "string"
					},
					"visibility": {
						"$ref": "#/components/schemas/ReagentVisibility"
					},
					"requesterOfferedReagentId": {
						"type": "string"
					},
					"restricted": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "enter 'Bearer {Firebase ID token}'"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/users/{id}": {
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "Update user information by their ID\r\nUser **must** be authenticated to access this endpoint",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- The ID of the user to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The user information to update",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_User_",
								"description": "- The user information to update"
							}
						}
					}
				}
			}
		},
		"/users/reagents/expiring": {
			"get": {
				"operationId": "GetReagentsExpiringSoon",
				"responses": {
					"200": {
						"description": "All reagents returned successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Reagent"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get reagents those are expiring in 60 days or less under a user by their user id",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/users/reagents": {
			"get": {
				"operationId": "GetReagents",
				"responses": {
					"200": {
						"description": "All reagents returned successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Reagent"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all reagents under a user\r\nUser must be authenticated to access this endpoint",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateReagent",
				"responses": {
					"201": {
						"description": "Reagent created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reagent"
								}
							}
						}
					}
				},
				"description": "Create a reagent by passing in all the required props.\r\nUser **must** be authenticated to access this endpoint (lab manager / admin) [seller]",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "- the request object containing the data of the reagent (id generated by db)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReagentRequest",
								"description": "- the request object containing the data of the reagent (id generated by db)"
							}
						}
					}
				}
			}
		},
		"/users/{user_id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "User retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/User"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "get user info using user id",
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"description": "- the id of the user to fetch",
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{user_id}/email": {
			"get": {
				"operationId": "GetEmail",
				"responses": {
					"200": {
						"description": "User email retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"email": {
											"type": "string"
										}
									},
									"required": [
										"email"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get email of the current logged in user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{user_id}/jwt": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "User information returned successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "Gets the basic information of a user",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- the ID of the user to retrieve",
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/reagents/{id}": {
			"get": {
				"operationId": "GetReagentById",
				"responses": {
					"200": {
						"description": "reagent returned successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Reagent"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Get a reagent by using its ID\r\nUser must be authenticated to access this endpoint",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- The ID of the reagent to fetch",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteReagentById",
				"responses": {
					"200": {
						"description": "reagent deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Reagent"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"description": "Delete a reagent by using its ID\r\nUser **must** be authenticated to access this endpoint",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- The ID of the reagent to delete",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateReagent",
				"responses": {
					"200": {
						"description": "Reagent updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reagent"
								}
							}
						}
					}
				},
				"description": "Update a reagent by its ID.\r\nUser **must** be authenticated to access this endpoint",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "- The ID of the reagent to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated reagent data (partial fields allowed)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Reagent_",
								"description": "- The updated reagent data (partial fields allowed)"
							}
						}
					}
				}
			}
		},
		"/users/{id}/reagents": {
			"get": {
				"operationId": "GetReagentsByUserId",
				"responses": {
					"200": {
						"description": "All reagents returned successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Reagent"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "\r\nGet all reagents under another user by their user id\r\nCan only be done by admin*",
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/reagents": {
			"get": {
				"operationId": "GetAllReagents",
				"responses": {
					"200": {
						"description": "Reagents retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Reagent"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all reagents with an option to filter them by category.",
				"security": [],
				"parameters": [
					{
						"description": "- The list of all categories to fetch reagents from.",
						"in": "query",
						"name": "category",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/components/schemas/ReagentCategory"
							}
						}
					}
				]
			},
			"post": {
				"operationId": "CreateReagent",
				"responses": {
					"201": {
						"description": "Reagent created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reagent"
								}
							}
						}
					}
				},
				"description": "Get a reagent by its ID.\nCan only be done by lab_admin and admin",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- the request object containing the data of the reagent (id generated by db)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReagentRequest",
								"description": "- the request object containing the data of the reagent (id generated by db)"
							}
						}
					}
				}
			}
		},
		"/reagents/{id}": {
			"get": {
				"operationId": "GetReagent",
				"responses": {
					"200": {
						"description": "Reagent retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reagent"
								}
							}
						}
					}
				},
				"description": "Get a reagent by its ID.",
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the reagent to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteReagent",
				"responses": {
					"200": {
						"description": "Reagent deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reagent"
								}
							}
						}
					}
				},
				"description": "Delete a reagent by its ID.\nCan only be done by lab_admin (who owns the reagent) and admin",
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the reagent to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateReagent",
				"responses": {
					"200": {
						"description": "Reagent updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Reagent"
								}
							}
						}
					}
				},
				"description": "Update a reagent by its ID.\nCan only be done by lab_admin (who owns the reagent) and admin",
				"security": [],
				"parameters": [
					{
						"description": "- The ID of the reagent to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The updated reagent data (partial fields allowed)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Reagent_",
								"description": "- The updated reagent data (partial fields allowed)"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}