# server/src/presentation-layer/controllers/UserController.ts
import { User } from "../../business-layer/models/User"
import {
  Controller,
  Get,
  Post,
  Delete,
  Route,
  SuccessResponse,
  Security,
  Request,
  Path,
  Body,
  Patch,
  Tags,
} from "tsoa"
import { UserService } from "../../data-layer/repositories/UserRepository"
import { Reagent } from "../../business-layer/models/Reagent"
import { ReagentService } from "../../data-layer/repositories/ReagentRepository"
import { CreateReagentRequest } from "../../service-layer/dtos/request/ReagentRequest"
import { AuthRequest } from "../../service-layer/dtos/request/AuthRequest"

@Tags("User")
@Route("users")
export class UserController extends Controller {
  /**
   * Update user information by their ID
   * User **must** be authenticated to access this endpoint
   * @param id - The ID of the user to update
   * @param user - The user information to update
   */
  @Security("jwt")
  @Patch("{id}")
  public async updateUser(
    @Path() id: string,
    @Body() user: Partial<User>,
  ): Promise<User> {
    try {
      console.log("UpdateUser called with:", { id, user })
      const updatedUser = await new UserService().updateUser(id, user)
      return updatedUser
    } catch (error) {
      console.error("Error updating user:", error)
      throw error
    }
  }

  /**
   * Get reagents those are expiring in 60 days or less under a user by their user id
   * @param user_id - user id to query with
   * @returns Promise<Reagent[]> - The list of all reagents filtered.
   */
  @SuccessResponse("200", "All reagents returned successfully")
  @Security("jwt")
  @Get("/reagents/expiring")
  public async getReagentsExpiringSoon(
    @Request() request: AuthRequest,
  ): Promise<Reagent[]> {
    const user_id = request.user?.uid
    if (!user_id) {
      this.setStatus(401)
      throw new Error("User ID not found")
    }

    return await new ReagentService().getReagentsExpiringSoon(user_id)
  }

  @SuccessResponse("200", "Users retrieved successfully")
  @Get()
  public async getAllUsers(): Promise<User[]> {
    console.log("Getting all users...")
    const users = await new UserService().getAllUsers()
    console.log(`Returning ${users.length} users`)
    return users
  }

  /**
   * Get all reagents under a user
   * User must be authenticated to access this endpoint
   * @param user_id - user id to query with
   * @returns Promise<Reagent[]> - The list of all reagents filtered.
   */
  @SuccessResponse("200", "All reagents returned successfully")
  @Security("jwt")
  @Get("/reagents")
  public async getReagents(
    @Request() request: AuthRequest,
  ): Promise<Reagent[]> {
    try {
      const user = request.user
      const user_id = user.uid
      if (user_id) {
        const reagents = await new ReagentService().getReagentsByUserId(user_id)
        return reagents
      }
      return []
    } catch (err) {
      throw new Error("Failed to fetch reagents: " + (err as Error).message)
    }
  }

  /**
   * get user info using user id
   *
   * @param user_id - the id of the user to fetch
   * @returns user info including display name
   */
  @SuccessResponse("200", "User retrieved successfully")
  @Get("{user_id}")
  public async getUserById(@Path() user_id: string): Promise<User | null> {
    const user = await new UserService().getUserById(user_id)
    return user
  }
  /**
   * Get email of the current logged in user
   *
   * @param jwt- jwt token is needed for the user to be verified
   * @returns email of the user who calls this endpoint
   */
  @Security("jwt")
  @SuccessResponse("200", "User email retrieved successfully")
  @Get("{user_id}/email")
  public async getEmail(
    @Request() request: AuthRequest,
    @Path() user_id: string,
  ): Promise<{ email: string }> {
    console.log("getting user's email..")
    console.log(user_id)
    const user = request.user
    return { email: user.email }
  }

  /**
   * Gets the basic information of a user
   *
   * @param jwt - jwt token is needed for the user to be verified
   * @param user_id - the ID of the user to retrieve
   * @returns Promise<User>
   */
  @Security("jwt")
  @SuccessResponse("200", "User information returned successfully")
  @Get("{user_id}/jwt")
  public async getUser(
    @Path() user_id: string,
    @Request() request: AuthRequest,
  ): Promise<User> {
    try {
      if (request.user.role !== "admin") {
        this.setStatus(403)
        console.error(
          "You don't have permission to access other users information",
        )
      }
      const user = await new UserService().getUserById(user_id)
      return user
    } catch (err) {
      throw new Error(`Failed to fetch user: ${(err as Error).message}`)
    }
  }

  @SuccessResponse("200", "retrieved user info")
  public async getUserF(@Path() id: string): Promise<User> {
    try {
      const user = await new UserService().getUserById(id)
      return user
    } catch (err) {
      throw new Error(`Failed to fetch user info: ${err as Error}`)
    }
  }

  /**
   * Get a reagent by using its ID
   * User must be authenticated to access this endpoint
   * @param id - The ID of the reagent to fetch
   * @returns Promise<Reagent> - object of type Reagent
   */
  @SuccessResponse("200", "reagent returned successfully")
  @Get("/reagents/{id}")
  @Security("jwt")
  public async getReagentById(@Path() id: string): Promise<Reagent | null> {
    try {
      const reagent = await new ReagentService().getReagentById(id)
      return reagent
    } catch (err) {
      throw new Error(
        `Failed to fetch reagent -- id:${id} ` + (err as Error).message,
      )
    }
  }

  /**
   * Delete a reagent by using its ID
   * User **must** be authenticated to access this endpoint
   * @param id - The ID of the reagent to delete
   * @returns Promise<Reagent> - object of type Reagent (deleted ref)
   */
  @SuccessResponse("200", "reagent deleted successfully")
  @Security("jwt")
  @Delete("/reagents/{id}")
  public async deleteReagentById(
    @Path() id: string,
    @Request() request: AuthRequest,
  ): Promise<Reagent | null> {
    try {
      const user_id = request.user.uid
      const reagent = await new ReagentService().getReagentById(id)
      if (user_id !== reagent.user_id) {
        console.error("You cannot delete a reagent that you don't own")
        return null
      }
      const reagentToDelete = await new ReagentService().deleteReagent(id)
      return reagentToDelete
    } catch (err) {
      throw new Error(
        `Failed to delete reagent -- id:${id} ` + (err as Error).message,
      )
    }
  }

  /**
   * Create a reagent by passing in all the required props.
   * User **must** be authenticated to access this endpoint (lab manager / admin) [seller]
   * @param id - The ID of the new reagent being created.
   * @param requestObject - the request object containing the data of the reagent (id generated by db)
   * @returns Promise<Reagent> - object of type Reagent (created ref)
   */
  @SuccessResponse("201", "Reagent created successfully")
  @Security("jwt")
  @Post("/reagents")
  public async createReagent(
    @Request() request: AuthRequest,
    @Body() requestObject: CreateReagentRequest,
  ): Promise<Reagent> {
    const user = request.user
    if (!user || !["admin", "lab_manager"].includes(user.role)) {
      console.error("You don't have permission to create reagents")
      throw new Error("Forbidden")
    }
    const user_id = request.user.uid
    const data = {
      ...requestObject,
      user_id,
    }
    const newReagent = await new ReagentService().createReagent(data as Reagent)
    return newReagent
  }
  /**
   * Update a reagent by its ID.
   * User **must** be authenticated to access this endpoint
   * @param id - The ID of the reagent to update.
   * @param update - The updated reagent data (partial fields allowed)
   * @returns Promise<Reagent> - object of type Reagent (updated ref)
   * @throws 400 - If the operation fails
   */
  @SuccessResponse("200", "Reagent updated successfully")
  @Security("jwt")
  @Patch("/reagents/{id}")
  public async updateReagent(
    @Path() id: string,
    @Body() update: Partial<Reagent>,
    @Request() request: AuthRequest,
  ): Promise<Reagent> {
    const uid = request.user?.uid
    if (!uid) {
      this.setStatus(401)
      throw new Error("User ID not authenticated")
    }
    try {
      const reagent = await new ReagentService().getReagentById(id)
      console.log(`User requesting update: ${uid}`)
      if (!reagent) {
        this.setStatus(404)
        throw new Error(`Reagent ID ${id} not found`)
      }

      if (reagent.user_id !== uid) {
        this.setStatus(403)
        throw new Error("Forbidden: You cannot update a reagent you do not own")
      }

      console.log(`User ${uid} is updating reagent ${id}`)
      const updatedReagent = await new ReagentService().updateReagent(
        id,
        update,
      )
      this.setStatus(200)
      return updatedReagent
    } catch (error) {
      if (!this.getStatus()) {
        this.setStatus(400)
      }
      throw new Error(`[ERROR] while updating reagent ${id}: ${error}`)
    }
  }

  /**
 
Get all reagents under another user by their user id
Can only be done by admin*
@param user_id - user id to query with
@returns Promise<Reagent[]> - The list of all reagents filtered.*/

  @SuccessResponse("200", "All reagents returned successfully")
  @Security("jwt")
  @Get("/{id}/reagents")
  public async getReagentsByUserId(@Path() id: string): Promise<Reagent[]> {
    try {
      const reagents = await new ReagentService().getReagentsByUserId(id)
      return reagents
    } catch (err) {
      throw new Error("Failed to fetch reagents: " + (err as Error).message)
    }
  }
}